LibGUI: Calculate unclamped_scrubber_size() as float
Large enough content ranges produced unclamped scrubbers sized zero,
effectively clamped by their integer type. This led to zero sized
page_increments and scrubbers which didn't budge on gutter events.
This fixes broken gutters in FontEditor and TextEditor for large
files.

+++ Userland/Libraries/LibGUI/Scrollbar.h
@@ -48,7 +48,7 @@ private:
     Gfx::IntRect decrement_button_rect() const;
     Gfx::IntRect increment_button_rect() const;
     Gfx::IntRect scrubber_rect() const;
-    int unclamped_scrubber_size() const;
+    float unclamped_scrubber_size() const;
     int visible_scrubber_size() const;
     int scrubbable_range_in_pixels() const;
     void on_automatic_scrolling_timer_fired();
